verify that ansible runner requirements are met (to fail early

+++ vagrant/run.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/bash
 
+python -c 'import netaddr'
+if [[ "$?" -ne 0 ]]; then
+ echo "You must install the python netaddr module"
+ exit 1
+fi
+
 set -e
 vagrant destroy
 vagrant up --provision
